requirements.txt
requests==2.18.4
scapy-python3==0.23